+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Major version bump due to dependencies
 - Updated `stylelint` to v14.0.0
+- Add stylelint-config-standard-scss
+- Disallow usage of `@extend` and `@import` statements
+- Ensure variables and `@include` statements are declared at the top of declaration blocks.
 
 # 4.0.0 - 2021-02-22